+++ api/records.go
@@ -37,8 +37,8 @@ func (api API) GetRecordsForPost(ctx context.Context, postID uint32) (*RecordsRe
 }
 
 // GetRecordsByID holds the business logic around getting many Records
-func (api API) GetRecordsByID(ctx context.Context, ids []uint32) ([]model.Record, error) {
-	records, err := api.recordPersister.SelectRecordsByID(ctx, ids)
+func (api API) GetRecordsByID(ctx context.Context, ids []uint32, enforceContextSocietyMatch bool) ([]model.Record, error) {
+	records, err := api.recordPersister.SelectRecordsByID(ctx, ids, enforceContextSocietyMatch)
 	if err != nil {
 		return nil, NewError(err)
 	}
@@ -86,7 +86,7 @@ func (api API) GetRecord(ctx context.Context, includeDetails bool, id uint32) (*
 		if err != nil {
 			return nil, NewError(err)
 		}
-		memberRecords, err := api.recordPersister.SelectRecordsByID(ctx, household.Records)
+		memberRecords, err := api.recordPersister.SelectRecordsByID(ctx, household.Records, true)
 		if err != nil {
 			return nil, NewError(err)
 		}
