+++ api/search_internal_test.go
@@ -272,6 +272,8 @@ func doInternalSearchTests(t *testing.T,
 	nameP model.NamePersister,
 ) {
 	ctx := utils.AddSocietyIDToContext(context.TODO(), 1)
+	ctx = utils.AddSearchUserIDToContext(ctx, 1)
+
 	testApi, err := NewAPI()
 	assert.NoError(t, err)
 	defer testApi.Close()
@@ -283,6 +285,7 @@ func doInternalSearchTests(t *testing.T,
 	}{
 		{
 			req: SearchRequest{
+				SocietyID:        1,
 				Given:            "Fred",
 				Surname:          "Flintstone",
 				SurnameFuzziness: FuzzyNameExact,
@@ -300,10 +303,13 @@ func doInternalSearchTests(t *testing.T,
                       	  {"match":{"given":{"query":"F","boost":0.2}}}
                     	]}}
 					  ]}}
-					]}},"from":0,"size":10}`,
+					],
+                    "filter":[{"term":{"societyId":{"value":1}}}]
+					}},"from":0,"size":10}`,
 		},
 		{
 			req: SearchRequest{
+				SocietyID:        1,
 				Given:            "Fred",
 				GivenFuzziness:   FuzzyNameExact | FuzzyNameVariants,
 				Surname:          "Flintstone",
@@ -317,10 +323,13 @@ func doInternalSearchTests(t *testing.T,
                     	]}},
 						{"match":{"surname":{"query":"Flintstone","boost":1}}}
 					  ]}}
-					]}},"from":0,"size":10}`,
+					],
+                    "filter":[{"term":{"societyId":{"value":1}}}]
+					}},"from":0,"size":10}`,
 		},
 		{
 			req: SearchRequest{
+				SocietyID:        1,
 				Given:            "Fred",
 				GivenFuzziness:   FuzzyNameExact | FuzzyNameSoundsLikeNarrow,
 				Surname:          "Flintstone",
@@ -334,10 +343,13 @@ func doInternalSearchTests(t *testing.T,
                     	]}},
 						{"match":{"surname":{"query":"Flintstone","boost":1}}}
 					  ]}}
-					]}},"from":0,"size":10}`,
+					],
+                    "filter":[{"term":{"societyId":{"value":1}}}]
+					}},"from":0,"size":10}`,
 		},
 		{
 			req: SearchRequest{
+				SocietyID:          1,
 				Surname:            "Flintstone",
 				SurnameFuzziness:   FuzzyNameExact,
 				BirthDate:          "1900",
@@ -357,10 +369,13 @@ func doInternalSearchTests(t *testing.T,
 						{"term":{"deathYear":{"value":"1995","boost":0.7}}},
 						{"range":{"deathYear":{"gte":1990,"lte":2000,"boost":0.3}}}
 					  ]}}
-					]}},"from":0,"size":10}`,
+					],
+                    "filter":[{"term":{"societyId":{"value":1}}}]
+					}},"from":0,"size":10}`,
 		},
 		{
 			req: SearchRequest{
+				SocietyID:          1,
 				Surname:            "Flintstone",
 				SurnameFuzziness:   FuzzyNameExact,
 				BirthDate:          "1900",
@@ -380,10 +395,13 @@ func doInternalSearchTests(t *testing.T,
 						{"term":{"deathYear":{"value":"1995","boost":0.7}}},
 						{"range":{"deathYear":{"gte":1990,"lte":2000,"boost":0.3}}}
 					  ]}}
-					]}},"from":0,"size":10}`,
+					],
+                    "filter":[{"term":{"societyId":{"value":1}}}]
+					}},"from":0,"size":10}`,
 		},
 		{
 			req: SearchRequest{
+				SocietyID:        1,
 				Surname:          "Flintstone",
 				SurnameFuzziness: FuzzyNameExact,
 				BirthPlace:       "Autauga, Alabama, United States",
@@ -398,10 +416,13 @@ func doInternalSearchTests(t *testing.T,
 						{"term":{"birthPlace3":{"value":"United States,Alabama,Autauga,","boost":1.0}}},
 						{"term":{"birthPlace2":{"value":"United States,Alabama","boost":0.4}}}
 					  ]}}
-					]}},"from":0,"size":10}`,
+					],
+                    "filter":[{"term":{"societyId":{"value":1}}}]
+					}},"from":0,"size":10}`,
 		},
 		{
 			req: SearchRequest{
+				SocietyID:           1,
 				Surname:             "Flintstone",
 				SurnameFuzziness:    FuzzyNameExact,
 				BirthPlace:          "Autauga, Alabama, United States",
@@ -415,10 +436,13 @@ func doInternalSearchTests(t *testing.T,
 						{"term":{"birthPlace3":{"value":"United States,Alabama,Autauga","boost":1.0}}},
 						{"term":{"birthPlace3":{"value":"United States,Alabama,Autauga,","boost":1.0}}}
 					  ]}}
-					]}},"from":0,"size":10}`,
+					],
+                    "filter":[{"term":{"societyId":{"value":1}}}]
+					}},"from":0,"size":10}`,
 		},
 		{
 			req: SearchRequest{
+				SocietyID:             1,
 				Surname:               "Flintstone",
 				SurnameFuzziness:      FuzzyNameExact,
 				CollectionPlace1:      "United States",
@@ -429,7 +453,7 @@ func doInternalSearchTests(t *testing.T,
 						{"match":{"surname":{"query":"Flintstone","boost":1}}}
 					  ]}}
 					],
-                    "filter":[{"term":{"collectionPlace1":{"value":"United States"}}}]
+                    "filter":[{"term":{"societyId":{"value":1}}},{"term":{"collectionPlace1":{"value":"United States"}}}]
 					}},
 					"aggs":{"collectionPlace2":{"terms":{"field":"collectionPlace2","size":250}}},
 					"from":0,"size":10}`,
