+++ client/src/views/Home.vue
@@ -47,7 +47,7 @@
     <v-footer paddless fixed style="background-color: #fafafa">
       <v-col cols="12">
         <a style="padding-right: 8px;" href="/privacy.html">Privacy policy</a>
-        <a style="padding-right: 8px;" href="/terms.html">Terms and conditions</a>
+        <a style="padding-right: 8px;" href="/terms.html">Terms of service</a>
         <a style="padding-right: 8px;" href="/cookie.html">Cookie policy</a>
       </v-col>
     </v-footer>
