+++ db/test_data/society.tsv
@@ -1,1 +1,1 @@
-{"name":"sandbox","secretKey":"SecretKey","loginURL":"https://ourroots.org","postMetadata":[]}
+{"name":"sandbox","secretKey":"MySecretKey","loginURL":"https://ourroots.org","postMetadata":[]}
