+++ model/record.go
@@ -14,7 +14,7 @@ import (
 // RecordPersister defines methods needed to persist records
 type RecordPersister interface {
 	SelectRecordsForPost(ctx context.Context, postID uint32) ([]Record, error)
-	SelectRecordsByID(ctx context.Context, ids []uint32) ([]Record, error)
+	SelectRecordsByID(ctx context.Context, ids []uint32, enforceContextSocietyMatch bool) ([]Record, error)
 	SelectOneRecord(ctx context.Context, id uint32) (*Record, error)
 	InsertRecord(ctx context.Context, in RecordIn) (*Record, error)
 	UpdateRecord(ctx context.Context, id uint32, in Record) (*Record, error)
