+++ search-client/.env.development
@@ -1,15 +1,1 @@
 VUE_APP_API_BASE_URL=http://localhost:8000
-
-VUE_APP_AUTH_PROVIDER = auth0
-VUE_APP_AUTH_DOMAIN = "https://ourroots.auth0.com"
-VUE_APP_AUTH_ENDPOINT_DOMAIN = "https://ourroots.auth0.com"
-VUE_APP_AUTH_CLIENT_ID = "Cb6Dc26RH5kB5wmdSUG8TPO3BwMifmod"
-VUE_APP_AUTH_REDIRECT_URL = "http://localhost:8080/callback.html"
-VUE_APP_AUTH_POST_LOGOUT_REDIRECT_URL = "http://localhost:8080"
-
-#VUE_APP_AUTH_PROVIDER = cognito
-#VUE_APP_AUTH_DOMAIN = "https://cognito-idp.us-east-1.amazonaws.com/us-east-1_Ueg25pMGY"
-#VUE_APP_AUTH_ENDPOINT_DOMAIN = "https://gsmd-test.auth.us-east-1.amazoncognito.com"
-#VUE_APP_AUTH_CLIENT_ID = "61n9ggblojaf6tfj1n4lre40g"
-#VUE_APP_AUTH_REDIRECT_URL = "http://localhost:8080/callback.html"
-#VUE_APP_AUTH_POST_LOGOUT_REDIRECT_URL = "http://localhost:8080"
