+++ search-client/public/index.html
@@ -13,6 +13,13 @@
     <noscript>
       <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
     </noscript>
+    <script>
+        window.ourroots = {
+          societyId: 1,
+          jwt: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxXzEiLCJleHAiOjE3MTQ0NDczOTN9._GpM2kJkxWwtixGO8kpaHpTfz6wyg3HJI815bioLgL0",
+          images_directory: "/"
+        };
+    </script>
     <div id="app"></div>
     <!-- built files will be auto injected -->
   </body>
