+++ utils/context_utils.go
@@ -9,11 +9,12 @@ import (
 
 const societyKey = "societyID"
 const userKey = "userKey"
+const searchUserKey = "searchUserID"
 
 func GetSocietyIDFromContext(ctx context.Context) (uint32, error) {
 	id, ok := ctx.Value(societyKey).(uint32)
 	if !ok {
-		return 0, errors.New("SocietyID not found in context")
+		return 0, errors.New("societyID not found in context")
 	}
 	return id, nil
 }
@@ -32,4 +33,16 @@ func GetUserFromContext(ctx context.Context) (*model.User, error) {
 
 func AddUserToContext(ctx context.Context, user *model.User) context.Context {
 	return context.WithValue(ctx, userKey, user)
+}
+
+func GetSearchUserIDFromContext(ctx context.Context) (uint32, error) {
+	id, ok := ctx.Value(searchUserKey).(uint32)
+	if !ok {
+		return 0, errors.New("search UserID not found in context")
+	}
+	return id, nil
+}
+
+func AddSearchUserIDToContext(ctx context.Context, userID uint32) context.Context {
+	return context.WithValue(ctx, searchUserKey, userID)
 }
